+++ maas/resource_maas_boot_resources_test.go
@@ -9,6 +9,7 @@ import (
 	"terraform-provider-maas/maas/testutils"
 	"testing"
 
+	"github.com/canonical/gomaasclient/client"
 	"github.com/canonical/gomaasclient/entity"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
@@ -48,18 +49,18 @@ func testAccMaasBootResources() string {
 data "maas_boot_source" "test" {}
 
 resource "maas_boot_source_selection" "mantic" {
-    boot_source = data.maas_boot_source.test.id
-    os 			= "ubuntu"
-    release		= "mantic"
-	arches     	= ["*"]
-	subarches  	= ["*"]
-	labels     	= ["*"]
+  boot_source = data.maas_boot_source.test.id
+  os 			= "ubuntu"
+  release		= "mantic"
+  arches     	= ["*"]
+  subarches  	= ["*"]
+  labels     	= ["*"]
 }
 
 resource "maas_boot_resources" "test" {
-    boot_source_selections = [
-        maas_boot_source_selection.mantic.id,
-    ]
+  boot_source_selections = [
+    maas_boot_source_selection.mantic.id,
+  ]
 }`
 }
 
@@ -188,4 +189,17 @@ func testAccCheckMaasBootResourcesDestroy(s *terraform.State) error {
 	}
 
 	return nil
+}
+
+func findBootSourceSelection(client *client.Client, boot_source int, os string, release string) (*entity.BootSourceSelection, error) {
+	if bootsourceselections, err := client.BootSourceSelections.Get(boot_source); err != nil {
+		return nil, err
+	} else {
+		for _, d := range bootsourceselections {
+			if d.OS == os && d.Release == release {
+				return &d, nil
+			}
+		}
+	}
+	return nil, nil
 }
